file order refactoring, thanks to red
