+++ app/assets/stylesheets/_bookings.scss
