+++ app/assets/stylesheets/_calendar.scss
