+++ app/assets/stylesheets/_carrousel.scss
