+++ app/assets/stylesheets/_modal.scss
