+++ app/assets/stylesheets/deviseLogin.scss
@@ -1,4 +1,3 @@
-@import '~bootstrap/scss/bootstrap.scss';
 @import './variables';
 
 body{
