+++ app/assets/stylesheets/main.scss
@@ -1,4 +1,3 @@
-@import '~bootstrap/scss/bootstrap.scss';
 @import './variables';
 @import './calendar';
 @import './carrousel';
