+++ app/javascript/packs/application.js
@@ -9,8 +9,8 @@
 //
 // const images = require.context('../images', true)
 // const imagePath = (name) => images(name, true)
-import '../stylesheets/deviseLogin.scss';
-import 'bootstrap';
+import 'bootstrap/dist/css/bootstrap.min';
+import '../../assets/stylesheets/deviseLogin.scss';
 
 require('@rails/ujs').start();
 require('turbolinks').start();
