+++ app/javascript/packs/index.js
@@ -1,5 +1,6 @@
-import '../stylesheets/main.scss';
-import 'bootstrap';
+import 'bootstrap/dist/css/bootstrap.min';
+import '../../assets/stylesheets/main.scss';
+
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { createStore } from 'redux';
