+++ app/javascript/stylesheets/application.scss
@@ -1,0 +1,1 @@
+@import '~bootstrap/scss/bootstrap.scss';
