+++ package.json
@@ -9,7 +9,7 @@
     "@rails/webpacker": "4.2.2",
     "axios": "^0.19.1",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
-    "bootstrap": "^4.4.1",
+    "bootstrap": "4.3.1",
     "eslint-plugin-import": "^2.19.1",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-react": "^7.17.0",
